requirement.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.1
django-tinymce==2.8.0
idna==2.8
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
